Added process tomography base class and started to consolidate code.
