+++ lightworks/tomography/__init__.py
@@ -19,8 +19,8 @@ Lightworks Tomography
 A set of tools for quantum state & process tomography on a system.
 """
 
-from .process_tomography import ProcessTomography
-from .process_tomography_mle import ProcessTomographyMLE
+from .process_tomography_li import LIProcessTomography
+from .process_tomography_mle import MLEProcessTomography
 from .state_tomography import StateTomography
 from .utils import (
     choi_from_unitary,
@@ -30,9 +30,9 @@ from .utils import (
 )
 
 __all__ = [
-    "ProcessTomography",
     "StateTomography",
-    "ProcessTomographyMLE",
+    "LIProcessTomography",
+    "MLEProcessTomography",
     "choi_from_unitary",
     "density_from_state",
     "process_fidelity",
