+++ tests/tomography/process_mle_test.py
@@ -15,7 +15,7 @@
 import pytest
 
 from lightworks import PostSelection, emulator, qubit
-from lightworks.tomography import ProcessTomographyMLE, choi_from_unitary
+from lightworks.tomography import MLEProcessTomography, choi_from_unitary
 
 
 def experiment(circuits, inputs, n_qubits):
@@ -42,9 +42,9 @@ cnot_exp = choi_from_unitary(
 )
 
 
-class TestProcessTomographyMLE:
+class TestMLEProcessTomography:
     """
-    Unit tests for ProcessTomographyMLE routine.
+    Unit tests for MLEProcessTomography routine.
     """
 
     def setup_class(self):
@@ -54,7 +54,7 @@ class TestProcessTomographyMLE:
         # Hadamard tomography
         n_qubits = 1
         circ = qubit.H()
-        self.h_tomo = ProcessTomographyMLE(
+        self.h_tomo = MLEProcessTomography(
             n_qubits, circ, experiment, [n_qubits]
         )
         self.h_tomo.process()
@@ -62,7 +62,7 @@ class TestProcessTomographyMLE:
         try:
             n_qubits = 2
             circ = qubit.CNOT()
-            cnot_tomo = ProcessTomographyMLE(
+            cnot_tomo = MLEProcessTomography(
                 n_qubits, circ, experiment, [n_qubits]
             )
             cnot_tomo.process()
